Refactored scan_astro-ph.py into multiple modules
Refactored all files and classes into more modules for better
modularity, and renamed some for clarity.
The operation flow stays the same.
