+++ scan_astroph/config.py
@@ -1,4 +1,7 @@
+"""Read configuration files"""
+
 import json
+
 
 def read_keywords(filename='keywords.txt'):
     with open(filename) as json_file:
