+++ setup.cfg
@@ -12,7 +12,7 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    scan_astro-ph = scan_astroph.scan_astroph:main
-    scan_astroph = scan_astroph.scan_astroph:main
-    astroph-scan = scan_astroph.scan_astroph:main
+    scan_astro-ph = scan_astroph.__main__:main
+    scan_astroph = scan_astroph.__main__:main
+    astroph-scan = scan_astroph.__main__:main
     scan_astro-ph.wordcounter = scan_astroph.wordcounter:main
